pyproject.toml
[tool.poetry]
name = "fill-habits"
version = "0.1.0"
description = ""
authors = [" annavaleriev <filenko.a@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.1.1"
djangorestframework = "^3.15.2"
psycopg2 = "^2.9.9"
psycopg2-binary = "^2.9.9"
django-cors-headers = "^4.4.0"
django-filter = "^24.3"
djangorestframework-simplejwt = "^5.3.1"
drf-spectacular = "^0.27.2"
python-dotenv = "^1.0.1"
coreapi = "^2.3.3"
celery = "^5.4.0"
redis = "^5.1.1"
eventlet = "^0.37.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
